requirements.txt
python-dotenv
requests
youtube-transcript-api
google-genai
google-api-python-client
google-auth-oauthlib
